requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
rembg==2.0.67
python-multipart==0.0.9
numpy==2.1.1
